[add] devicehelper
